+++ absl/logging/__init__.py
@@ -86,6 +86,7 @@ import os
 import socket
 import struct
 import sys
+import tempfile
 import threading
 import time
 import timeit
@@ -707,6 +708,9 @@ def find_log_dir(log_dir=None):
     OSError: raised in Python 2 when it cannot find a log directory.
   """
   # Get a list of possible log dirs (will try to use them in order).
+  # NOTE: Google's internal implementation has a special handling for Google
+  # machines, which uses a list of directories. Hence the following uses `dirs`
+  # instead of a single directory.
   if log_dir:
     # log_dir was explicitly specified as an arg, so use it and it alone.
     dirs = [log_dir]
@@ -715,7 +719,7 @@ def find_log_dir(log_dir=None):
     # behavior of the same flag in logging.cc).
     dirs = [FLAGS['log_dir'].value]
   else:
-    dirs = ['/tmp/', './']
+    dirs = [tempfile.gettempdir()]
 
   # Find the first usable log dir.
   for d in dirs: